==> Project/Nevus/component/1-descriptionFull.ipynb <==
import json
import os
import pandas
import tqdm
os.getcwd()

##  要補註解

path = "C:\\Users\\user\\Desktop\\Python\\projectNevus\\data\\description\\"
N = len(os.listdir(path))
Id        = [None] * N
Age       = [None] * N
Status    = [None] * N
Sex       = [None] * N
Diagnosis = [None] * N
Type      = [None] * N
for i, file in enumerate(os.listdir(path)):
    
    Id[i] = file  
    
    Json = json.loads(open(path + file).read())
    if( "meta" in Json.keys() ):
        if( "clinical" in Json["meta"].keys() ):
            if( "age_approx" in Json["meta"]["clinical"].keys() ):
                Age[i] = Json["meta"]["clinical"]["age_approx"]
            if( "benign_malignant" in Json["meta"]["clinical"].keys() ):
                Status[i] = Json["meta"]["clinical"]["benign_malignant"]
            if( "sex" in Json["meta"]["clinical"].keys() ):
                Sex[i] = Json["meta"]["clinical"]["sex"]
            if( "diagnosis" in Json["meta"]["clinical"].keys() ):
                Diagnosis[i] = Json["meta"]["clinical"]["diagnosis"]
        if( "acquisition" in Json["meta"].keys() ):
            if( "image_type" in Json["meta"]["acquisition"].keys() ):
                Type[i] = Json["meta"]["acquisition"]["image_type"]


descriptionFull = pandas.DataFrame({
    "Id" : Id,
    "Age" : Age,
    "Status": Status,
    "Sex": Sex,
    "Diagnosis": Diagnosis,
    "Type": Type
})
descriptionFull.head()

descriptionFull.shape

path = "C:\\Users\\user\\Desktop\\Python\\projectNevus\\data\\temporary\\"
descriptionFull.to_csv(path + "descriptionFull.csv")
