==> Project/Nevus/component/2-descriptionClear.ipynb <==
import os 
import pandas
import numpy
os.getcwd()

##  要補註解

path = "C:\\Users\\user\\Desktop\\Python\\projectNevus\\data\\temporary\\"
file = "descriptionFull.csv"
descriptionFull = pandas.read_csv(path + file, index_col = 0)
descriptionFull.head()

##  刪除有遺失值的資料
descriptionClear = descriptionFull.dropna()
descriptionClear.shape

file = "descriptionClear.csv"
descriptionClear.to_csv(path + file)
