==> Project/Nevus/component/3-descriptionUseful.ipynb <==
import os 
import pandas
import numpy
import pickle
os.getcwd()

##  補註解

path = "C:\\Users\\user\\Desktop\\Python\\projectNevus\\data\\temporary\\"
file = "descriptionClear.csv"
descriptionClear = pandas.read_csv(path + file, index_col=0)
descriptionClear.head()

##  痣的診斷
pandas.get_dummies(descriptionClear.Diagnosis).sum(axis=0)

##  痣的狀態
pandas.get_dummies(descriptionClear.Status).sum(axis=0)

##  補註解

descriptionUseful  = descriptionClear

index = []
for i in list(descriptionUseful.Diagnosis.values):
    if(i=="melanoma" or i=="nevus"):
        index.append(True)
    else: 
        index.append(False)

descriptionUseful = descriptionUseful[index]

index = []
for i in list(descriptionUseful.Status.values):
    if(i=="benign" or i=="malignant"):
        index.append(True)
    else: 
        index.append(False)

descriptionUseful = descriptionUseful[index]

Status = pandas.get_dummies(descriptionUseful.Status).sum(axis=0)
Status/sum(Status)

path = "C:\\Users\\user\\Desktop\\Python\\projectNevus\\data\\temporary\\"
descriptionUseful.to_csv(path + "descriptionUseful.csv")
