==> Project/Nevus/component/4-dataInputFull.ipynb <==
import os 
import PIL
import numpy
import pandas
import tqdm
import pickle
os.getcwd()

path = "C:\\Users\\user\\Desktop\\Python\\projectNevus\\data\\temporary\\"
file = "descriptionUseful.csv"
descriptionUseful = pandas.read_csv(path+file, index_col=0)
descriptionUseful.head()

Sex    = pandas.get_dummies(descriptionUseful.Sex)
Status = pandas.get_dummies(descriptionUseful.Status)
descriptionUseful = pandas.concat([descriptionUseful, Sex], axis=1)
descriptionUseful.head()

##  解決 PIL 開啟圖檔出現 error 的問題
from PIL import ImageFile
ImageFile.LOAD_TRUNCATED_IMAGES = True 

##  解決 tqdm_notebook 出現 error 問題
tqdm.tqdm.monitor_interval = 0

path = "D:\\ISIC-Archive-Downloader\\Data\\Images\\"

Id       = []
Image    = []
Variable = []
Label    = []
Index    = []
for i in tqdm.tqdm_notebook(descriptionUseful.Id):
    
    file   = i + ".jpg"
    resize = (32,32)
    
    if( file in os.listdir(path) ):
        
        image    = PIL.Image.open(path + file)
        image    = image.resize(resize, PIL.Image.BILINEAR)
        variable = descriptionUseful[descriptionUseful.Id == i][["Age","female","male"]].values
        status   = descriptionUseful[descriptionUseful.Id == i].Status.values

        if status == 'benign':
            label = 1
        elif status == 'malignant':
            label = 0
        else:
            label = None
        
        if( numpy.array(image).shape[2] == 3 ):           
            Id.append(numpy.array(i))
            Image.append(numpy.array(image))
            Variable.append(numpy.array(variable))
            Label.append(numpy.array(label))

Index    = numpy.array([i for i in range(len(Id))])
Id       = numpy.array(Id)
Image    = numpy.array(Image)
Variable = numpy.array(Variable)  ##  這邊 dimension 會多 1
Variable = Variable.reshape((Variable.shape[0],Variable.shape[2]))  ##  修正 dimension
Label    = numpy.array(Label)
dataInputFull = (Index, Id, Image, Variable, Label)

path = "C:\\Users\\user\\Desktop\\Python\\projectNevus\\data\\temporary\\"
file = "dataInputFull.txt"
pickle.dump(dataInputFull, open(path + file, 'wb'))

# loadTest = pickle.load(open(path + file, 'rb'))
# print(type(loadTest),  "\n")
# print(loadTest[0][1:6],"\n")
# print(loadTest[2][1:6],"\n")
