==> Project/Nevus/component/6-Kfold.ipynb <==
import os 
import pickle
import numpy

path      = "C:\\Users\\user\\Desktop\\Python\\projectNevus\\data\\temporary\\"
file      = "dataInput.txt"
dataInput = pickle.load(open(path + file, "rb"))

from sklearn.model_selection import StratifiedKFold

K     = 4
Id    = dataInput[0]
Label = dataInput[3]

for indexTrain, indexValid in StratifiedKFold(n_splits=K,random_state=0,shuffle=True).split(Id , Label):
    print('Train: %s | test: %s' % (indexTrain, indexTrain))
